# feat: reduce memory allocation for common slice and map types