--- a/ldvalue/value_array.go
+++ b/ldvalue/value_array.go
@@ -248,3 +248,116 @@ func (a ValueArray) Transform(fn func(index int, value Value) (Value, bool)) Val
 func (a ValueArray) String() string {
 	return a.JSONString()
 }
+
+func copyArbitraryArrayString(o []string) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = String(v)
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayBool(o []bool) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Bool(v)
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayInt(o []int) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayInt8(o []int8) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayInt16(o []int16) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayInt32(o []int32) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayInt64(o []int64) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayUint(o []uint) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayUint8(o []uint8) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayUint16(o []uint16) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayUint32(o []uint32) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayUint64(o []uint64) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayFloat32(o []float32) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(float64(v))
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}
+func copyArbitraryArrayFloat64(o []float64) Value {
+	a := make([]Value, len(o))
+	for i, v := range o {
+		a[i] = Float64(v)
+	}
+
+	return Value{valueType: ArrayType, arrayValue: ValueArray{data: a}}
+}