--- a/ldvalue/value_map.go
+++ b/ldvalue/value_map.go
@@ -274,6 +274,7 @@ func copyArbitraryMapString(data map[string]string) Value {
 	for k, v := range data {
 		m[k] = String(v)
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -282,6 +283,7 @@ func copyArbitraryMapBool(data map[string]bool) Value {
 	for k, v := range data {
 		m[k] = Bool(v)
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -290,6 +292,7 @@ func copyArbitraryMapInt(data map[string]int) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -298,6 +301,7 @@ func copyArbitraryMapInt8(data map[string]int8) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -306,6 +310,7 @@ func copyArbitraryMapInt16(data map[string]int16) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -314,6 +319,7 @@ func copyArbitraryMapInt32(data map[string]int32) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -322,6 +328,7 @@ func copyArbitraryMapInt64(data map[string]int64) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -330,6 +337,7 @@ func copyArbitraryMapUint(data map[string]uint) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -338,6 +346,7 @@ func copyArbitraryMapUint8(data map[string]uint8) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -346,6 +355,7 @@ func copyArbitraryMapUint16(data map[string]uint16) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -354,6 +364,7 @@ func copyArbitraryMapUint32(data map[string]uint32) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -362,6 +373,7 @@ func copyArbitraryMapUint64(data map[string]uint64) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
@@ -370,13 +382,15 @@ func copyArbitraryMapFloat32(data map[string]float32) Value {
 	for k, v := range data {
 		m[k] = Float64(float64(v))
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }
 
 func copyArbitraryMapFloat64(data map[string]float64) Value {
 	m := make(map[string]Value, len(data))
 	for k, v := range data {
-		m[k] = Float64(float64(v))
+		m[k] = Float64(v)
 	}
+
 	return Value{valueType: ObjectType, objectValue: ValueMap{data: m}}
 }