@cwaldren-ld 2024-09-10
Hi @dyolcekaj, thank you for the detailed description and patch. It's appreciated.

At a first glance, the added verbosity seems to be worth the improvement. I'll give it a more thorough review as soon as I can. 

_Filed internally as 255748_.