@dyolcekaj 2024-10-17
> Hi @dyolcekaj, thank you for the detailed description and patch. It's appreciated.
> 
> At a first glance, the added verbosity seems to be worth the improvement. I'll give it a more thorough review as soon as I can.
> 
> _Filed internally as 255748_.

Hi @cwaldren-ld, 

Thanks again for considering this change.  Just following up on any internal discussions, curious to know if any decisions on accepting or rejecting this change have been made. 

Thanks!