@cwaldren-ld 2024-10-21
Hi @dyolcekaj , thanks again for this. It went out as [3.2.0](https://github.com/launchdarkly/go-sdk-common/releases/tag/v3.2.0). This will be pulled into the Go SDK shortly (hopefully.)